map added to ricerca page
